Create README.md
At the beginning of every financial year, every employee submits a list of potential declarations for that year. These declaration values are compared with their corresponding exemption limits and appropriate value is deducted from the employee's gross salary. This value obtained after deduction is called as the taxable income of that employee. Tax is calculated on this taxable income on the basis of which slab it belongs to. Every slab has a percentage value associated with it. Using these values, the tax is calculated.
This software system allows an employee to add/edit/remove declaration values at any point of time in the year. As soon as a modification is made, the taxable income and income tax of that employee gets updated for the remaining months of the year. Similarly, the system is capable of accommodating changes in declaration fields, exemption limits, tax slabs.